integrate new fields with multitrack
